refactor: Extract digit object
