+++ src/main/java/calculator/Calculator.java
@@ -6,7 +6,6 @@ import calculator.calculatableImpl.Multiplier;
 import calculator.calculatableImpl.Subtractor;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 public enum Calculator {
     ADDER(new Adder()),
@@ -15,7 +14,6 @@ public enum Calculator {
     DIVIDER(new Divider());
 
     private final static String INPUT_REGEX = " ";
-    private final static String DIGIT_REGEX = "^[0-9]*$";
     private final static String NO_INPUT_EXCEPTION = "입력 값이 null 이거나 빈 공백 문자입니다.";
     private final Calculatable calculatable;
 
@@ -26,8 +24,8 @@ public enum Calculator {
     public static int run(String inputString) {
         validateInput(inputString);
         List<String> inputs = List.of(inputString.split(INPUT_REGEX));
-        List<String> operators = Operation.getOperators(inputs);
-        List<Integer> digits = extractDigits(inputs);
+        List<String> operators = Operation.getFrom(inputs);
+        List<Integer> digits = Digit.parseDigits(inputs);
         int result = digits.get(0);
         for (int i = 0; i < operators.size(); i++) {
             int nextOperand = digits.get(i + 1);
@@ -61,16 +59,5 @@ public enum Calculator {
         }
 
         return DIVIDER.calculatable;
-    }
-
-    private static List<Integer> extractDigits(List<String> inputs) {
-        return inputs.stream()
-                .filter(Calculator::isDigit)
-                .map(Integer::parseInt)
-                .collect(Collectors.toList());
-    }
-
-    public static boolean isDigit(String inputString) {
-        return inputString.matches(DIGIT_REGEX);
     }
 }
